- use array precision if input is array
- avoid flot to string conversion with `str` as this might use `decimal.ROUND_HALF_UP` depending on the python version
- speedup writing a large number of rows:

```python
from mrtrix3.io import image
import numpy as np
im = image.Image()
im.data = np.zeros((10,11,12))
np.random.seed(0)
im.grad = np.random.rand(5000,4)
im.transform=0.1 * np.random.rand(4,4) + np.eye(4)
im.comments = ['comment ' + str(i) for i in range(5000)]

%timeit im.save('test.mif')
```
```
old: 40.3 ms ± 1.91 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
new: 13.2 ms ± 701 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

Notable differences of the new version compared to the current version are changes to the output precision and the added fourth `transform` line. The latter currently does not matter as the last row is not used in MRtrix3 so might just as well ditch it but up to you. 

```
head old.mif 
mrtrix image
dim: 10,11,12
vox: 1.0,1.0,1.0
layout: +2,+1,+0
datatype: Float64LE
transform: 1.0392172961350865,0.004115658616429163,0.09233005685513089,0.04062349745714864
transform: 0.09442821831399317,1.0722724489680333,0.09183199522385971,0.08232675386765655
transform: 0.0646896701857699,0.006793902002339891,1.083881577205911,0.0692822400555904
comments: comment 0
comments: comment 1
```
```
head new.mif
mrtrix image
dim: 10,11,12
vox: 1,1,1
layout: +2,+1,+0
datatype: Float64LE
transform: 1.03921729613509,0.00411565861642916,0.0923300568551309,0.0406234974571486
transform: 0.0944282183139932,1.07227244896803,0.0918319952238597,0.0823267538676565
transform: 0.0646896701857699,0.00679390200233989,1.08388157720591,0.0692822400555904
transform: 0.0957991514481232,0.0142222468130509,0.00877735378404738,1.05188799807261
comments: comment 0
```