@dchristiaens 2020-11-20
Thanks Max. All looks fine, except I'd prefer the transform to omit the last line and be consistent with the default output headers in MRtrix (e.g. converting dicom to .mif writes only 3 lines. I'm merging this and will change this after.